Applying changes from user prompt: Could you move the code for spliceFileWithOriginal out of assistant.go and into parseresponse.go?
