+++ assistant/chatgpt/chatgpt.go
@@ -110,4 +110,3 @@ func SendRequest(request ChatGPTRequest) (string, error) {
 
     return "", fmt.Errorf("no response from ChatGPT")
 }
-
